feat: add face culling management to webgl state and allow layers to indicate culling mode (#526)
### Summary
<!---
  Please try to cover the following questions in your summary:
  1. Describe the product (or technical) problem you are solving.
  2. Explain the engineering solution you have chosen.
  3. Discuss the implementation choices/tradeoffs you have made.
-->

Primary goal - allow the volume rendering layer to cull back faces
during rendering.
Adds to the webgl state to allow setting a cullFace mode. Primarily for
culling back faces, but allows all modes (front, back, front and back)
or being disabled. Follows a similar pattern to the blending mode in the
webgl state abstraction.

The example of volume rendering looks a bit odd right now with this, but
I'm not entirely sure why as the state abstraction seems correctly
setup. To debug this, I tried creating a simpler box geometry, similar
to neuroglancer style
https://github.com/google/neuroglancer/blob/master/src/webgl/bounding_box.ts
-- this one is actually in glsl, but it's the same idea of hardcoding
all the faces, normals, and uvs in one vertex array and hardcoding the
indices in another array, as opposed to the face building operation
which the box_geometry uses. The back face culling seems fine with the
simple geometry https://github.com/chanzuckerberg/idetik/pull/515

Since the scope of this PR is to setup how layers could indicate to the
rendering wanting to use face culling, that box geometry issue could
possibly be solved separately.

### Tests & Checks
<!---
  How did you validate that your changes were implemented correctly?
  
  Please think about the following prompts:
  1. I wrote automated tests.
  2. I manually tested my changes, and here's what I did:
-->
Manually tested the examples