chore: configure dependabot to add a conventional commit prefix (#537)
### Summary
Every time dependabot makes a PR it fails the PR title test. I'm hoping
this will fix it.

I'm not sure it will take effect for the existing PRs, so we might need
to wait for the next to confirm it's working.

I also added configs to skip the title check for draft PRs just to show
a bit less red in the PR list.