chore: remove old example PNG (#541)
### Summary
This change removes a PNG that was used for one of our initial examples,
but is no longer used. I chose to remove it now as part of ongoing clean
up efforts. As this is a binary blob of a non-trivial size, we could
also consider removing this from the history of the repo, but it's
relatively small and probably not worth the effort.