feat: use factory methods instead of overloaded constructors for OmeZarrImageSource (#539)
### Summary
This changes the public way to create an `OmeZarrImageSource` from using
overloaded constructors to using factory methods. This improves
readability of the implementation and will likely improve the
readability of external usage, especially as we add more source types
(e.g. private s3 buckets in #534) and arguments.

### Related Issue
Closes #536

### Tests & Checks
I manually tested all the examples.

---------

Co-authored-by: phoenixAja <plogan@chanzuckerberg.com>