refactor: restrict s3 store types (#542)
This restricts the types used by the `S3FetchStore` in the
`OmeZarrImageSource` by limiting the associated properties to a
corresponding factory method `OmeZarrImageSource.fromS3` instead of
sharing that with `fromHttp` which may have nothing to do with AWS or
S3.

It also removes some of the associated exports in the main index for
similar reasons.