fix: refactor size observer for immediate resizing (#543)
Summary:
This PR fixes a glitch where the canvas would look stretched for a moment when
resizing the viewport. The PixelSizeObserver relied on the animation loop
calling getAndResetChanged() for resizing which introduced a one frame delay.
I added a callback instead of waiting for the next animation frame to mitigate it.