fix: Prevent unwanted disposal of fallback chunks and add sanity check (#546)
### Summary
This is intended to fix #538, unwanted disposal of fallback chunks (not
visible or prefetch, but assigned a priority) by disposing all chunks
with null priority. This also adds a logic sanity check and will throw
an error if any visible or prefetch chunks have null priority. See
`chunk_store.ts`: 202-219.

### Related Issue
Closes #538

### Tests & Checks
Tested all examples, in particular the OME-Zarr v0.5 example that was
exhibiting the bug