fix: correct winding order of triangles in box geometry on back faces
The check was for if the face was at the front or back to then pack the indices into the
index buffer differently. But because of the udir and vdir in the function meaning that
the positions of a b c and d on a face change anyway because of udir and vdir I don't
think the extra wdir if check is needed to change the order of the indices. I see from
the commit introducing the box geometry that it is based on threejs and I could be
missing something but I don't see the equivalent check on wdir to change the index
insertion order in that either.